src/test/java/playground/repository/UserRepositoryTest.java
package playground.repository;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.transaction.annotation.Transactional;
import playground.domain.User;

import static org.assertj.core.api.Assertions.assertThat;

@SpringBootTest
@Transactional
class UserRepositoryTest {

    @Autowired
    private UserRepository userRepository;

    @Test
    @DisplayName("회원을 저장한다.")
    void save() {
        //given
        User user = new User("a@naver.com", "password", "김성빈");

        //when
        long userId = userRepository.save(user);

        //then
        assertThat(userId).isNotZero();
    }
}
